Fixed issue with progress bar in horizontal container...
If you set a min_width on an element it got drawn with a default fraction of
the total width, rather than filling up all the space. I captured this in a
unit test and then set about redesigning HorizontalContainer.draw() AGAIN!

It's not pretty ugly, but it WORKS. I'm now going to document and refactor.
Yuck, some of this code is nasty. Who'd work with UIs eh?