Rearranged modules, added some Python logging setup...
Wanted to split up pyamp as it was not focussed enough in scope, but
therefore ended up having to fiddle with the late gst import again because
a teeny bit of gst usage has currently bled out of the player module.
Considering refactoring so that we can do away with any nastiness!

I've also made a PyampBase class, which simply makes a logger available for
each child class so that I don't have to manually keep getting loggers
everywhere.