Got all meta...
Reading the sqlite3 documentation, there was some interesting stuff about
how one can register classes as adapters/converters to and from SQL tables
(basically, the exact thing I was trying to solve nastily with lots of
string processing).

I implemented a class for TrackMetadata, but then decided that it would be
good to generalise it a) for practice and b) so that I could use it for
Dir. In doing so, I wanted to use an abstract base class, because they're
cool, which started me down the road of metaclasses.

In addition to providing the ABC functionality, my custom metaclass
registers each newly defined class with sqlite as a converter, and does
some munging of __slots__ to make them work, to save the class author having
to get either thing right.

I'm still not sure I like metaclasses, but this was VERY good practice, and
we're not planning to expose these classes to others, so we shouldn't run
the risk of confusing everyone!