Added Queue class, ancilliary stuff and tests...
The Queue class is actually quite smart - it combines auto-generating
modes of track listing with specific tracks queued by the user. This means
that a Queue.next() is actually a coroutine though, because we may have to
go off to the database and grab some track data. That makes testing a bit
of a pain :-( Can't think how to make it nicer right now, and also don't
like the asymmetry between next (unknown, so coroutine) and previous (just
jump back in the list, so regular callable).